Add InfoItem component
